# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ShapesDocumentInfo" : {
        "properties" : {
          "default" : {
            "type" : "boolean"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "order" : {
            "format" : "int32",
            "type" : "integer"
          },
          "origin" : {
            "enum" : [
              "IMPORTED",
              "AUTHORED"
            ],
            "type" : "string"
          },
          "sourceFileName" : {
            "type" : "string"
          },
          "tripleCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "contact" : {
      "name" : "soptim",
      "url" : "https://www.soptim.de/"
    },
    "description" : "This API provides utilities for editing RDFGraphs that model UML classes using the CIM standard.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "RDFArchitect backend",
    "version" : "0.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/datasets/{datasetName}/graphs/{graphURI}/shacl/documents/{documentId}" : {
      "delete" : {
        "description" : "Deletes a set of SHACL shapes. The graph's default document cannot be deleted, only emptied.",
        "operationId" : "deleteShapesDocument",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The id of the constraints document.",
            "in" : "path",
            "name" : "documentId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "delete a constraints document",
        "tags" : [
          "shacl"
        ]
      },
      "get" : {
        "description" : "Returns the document's Turtle source, as the user last saved it where that text is known.",
        "operationId" : "getShapesDocumentText",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The id of the constraints document.",
            "in" : "path",
            "name" : "documentId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "get constraints document content",
        "tags" : [
          "shacl"
        ]
      },
      "patch" : {
        "description" : "Renames a document, switches it on or off, or moves it in the list. Omitted parameters are left unchanged. Order affects list position and serialisation order only — SHACL has no precedence between documents.",
        "operationId" : "updateShapesDocument",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The id of the constraints document.",
            "in" : "path",
            "name" : "documentId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          },
          {
            "description" : "New display name.",
            "in" : "query",
            "name" : "name",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Whether the shapes take part in validation and combined export.",
            "in" : "query",
            "name" : "enabled",
            "required" : false,
            "schema" : {
              "type" : "boolean"
            }
          },
          {
            "description" : "New position in the document list.",
            "in" : "query",
            "name" : "order",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShapesDocumentInfo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "update constraints document metadata",
        "tags" : [
          "shacl"
        ]
      },
      "put" : {
        "description" : "Replaces the document's shapes with the given Turtle.",
        "operationId" : "replaceShapesDocumentText",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The id of the constraints document.",
            "in" : "path",
            "name" : "documentId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "text/plain" : {
              "schema" : {
                "description" : "The SHACL shapes in Turtle syntax.",
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "replace constraints document content",
        "tags" : [
          "shacl"
        ]
      }
    },
...
  }
}